nanny_bot/requirements.txt
fastapi
uvicorn[standard]
python-telegram-bot==21.*
SQLAlchemy>=2.0
asyncpg
python-dotenv
pydantic>=2.0